requirements.txt
azure-common==1.1.28
azure-core==1.24.0
azure-servicebus==7.6.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.2.0
uamqp==1.5.3
urllib3==1.26.9
